Add byte_array_to_charstr().
If the byte array is shared, this decrements the reference count and
returns a copy of the contents.  Otherwise, it returns the actual array,
trimmed down to its contents, and destroys the metadata structure.  In
both cases, the caller has full ownership of the returned string and is
responsible for freeing it.

This makes it possible to use a byte array to build a string, then
return the result without incurring an additional allocation and copy.